requirements.txt
anthropic>=0.25.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pypdf>=4.1.0
reportlab>=4.1.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pytest>=8.1.0
httpx>=0.27.0
